customer

+++ src/main/java/com/bookstore/service/CategoryServices.java
@@ -8,6 +8,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.bookstore.dao.BookDAO;
 import com.bookstore.dao.CategoryDAO;
 import com.bookstore.entity.Category;
 
@@ -97,9 +98,18 @@ public class CategoryServices {
 
 	public void deleteCategory() throws ServletException, IOException {
 		int categoryId = Integer.parseInt(request.getParameter("id"));
-		categoryDAO.delete(categoryId);
-		
-		String message = "The category with ID " + categoryId + " has been deleted successfully.";
-		listCategory(message);
-	}
+		BookDAO bookDAO= new BookDAO();
+	    long numberOfBooks=bookDAO.countByCategory(categoryId);
+	    String message;
+
+	    if (numberOfBooks > 0)  {
+	    	message= "Could not delete the category (ID: %d) because it contains some books.";
+	    	message = String.format(message,numberOfBooks);
+	    } 
+	    else {
+	    	categoryDAO.delete(categoryId);
+	    	message = "The category with ID " + categoryId + " has been removed successfully.";
+	    }
+	    listCategory(message);
+	  }
 }
